# Using the BIT instruction in Armv8 table selection.