### Description of changes: 
Table selection function select_w5 is 35% faster than before (AND/OR combination), making P-256
ECDH, ECDSA Sign and ECDSA Verify faster by
2%, 3.7% and 1%, respectively.

| P-256, GV2   | Before (Ops/s) | After (MB/s) | Improvement |
:--------------|---------------:|-------------:|:------------:
| ECDH         |      6019      |     6140     |    1.02x    |
| ECDSA Sign   |     19028      |    19732     |    1.037x   |
| ECDSA Verify |      6595      |     6656     |    1.01x    |

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.